mlModule/model_api/requirements.txt
absl-py==0.12.0
aniso8601==9.0.1
astunparse==1.6.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
cycler==0.10.0
Flask==2.0.0
Flask-RESTful==0.3.8
flatbuffers==1.12
gast==0.3.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
joblib==1.0.1
kafka-python==2.0.2
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.0
oauthlib==3.1.0
opt-einsum==3.3.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycoingecko==2.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
termcolor==1.1.0
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==2.0.0
wrapt==1.12.1
